requirements.txt
Pillow~=9.5.0
pip~=21.3.1
wheel~=0.37.1
numpy~=1.24.3
setuptools~=60.2.0
rawpy~=0.18.1
PyExifTool~=0.5.5